Added navigation drawer to MainActivity.java. The layout file activity_main_drawer_header.xml contains the navigation drawer header layout. The layout file activity_main_drawer_menu.xml contains a list of the options one can click in the navigation drawer.
Added TableListFragment.java and OrderQueueFragment.java, the first of
many fragments. The layout file fragment_table_list.xml and
fragment_order_queue.xml contain the layouts for each of these
fragments, respectively.

Each of these fragments is linked with the navigation drawer to appear
when they are summoned by the navigation drawer.