Organized Java classes into four categories: Activities, Adapters, Fragments, Models.
Activities and Fragments are self-explanatory. Adapters are classes that adapt data from raw form into data that can be fed into a ListView and displayed to the user. Models are classes that represent some aspect of our system. So we will eventually have a User class, a Table class, etc.