Completely implemented the ManageEmployees module of the application. Added activities include:
ManageEmployeeFragment: Displays a ListView of Employee objects stored in the database.
AddEmployeeActivity: Provides a form for the user to fill out to create a new Employee object in the database.
ModifyEmployeeActivity: Provides a form for the user to modify or delete an existing Employee object in the database.

Other small changes include: Organized strings inside of strings.xml values file, added more styles inside of styles.xml file.