Made string names and android resource id names more consistent.
String names will have a prefix followed by a short description describing the view's function that the string is attached to. So, for example, activity and fragment names have the prefix "name". Suppose we want to name a string resource for the MainActivity's activity name. The string resource name will have the name "name_main". Look thorough the way strings are grouped and familiarize yourself with the prefixes.

Android resource id's are similar, but slightly more complex. Their prefix will be the element they are, followed by the activity/fragment/item name they appear in, followed by a short description describing the view (if necessary). So, for example, suppose we are tasked with naming the android resource id of a Button in "activity_login.xml". This button logs a user in and is the only button in the layout. The android resource id would be named "button_login". A short description is not necessary in this case, because it is the only button on the layout.