Completely implemented the ManageMenu module. Added Java classes include:
AddMenuItemActivity.java: Provides the user with a form to add a new MenuItem to the database.
ModifyMenuItemActivity.java: Provides the user with a form to modify or delete an existing MenuItem from the database.
MenuItemsBaseAdapter.java: Defines a BaseExpandableListAdapter for the ExpandableListView in the ManageMenuFragment.
MenuItemsValueEventListener: Defines a ValueEventListener used in the ManageMenuFragment.
ManageMenuFragment: Provides the user with an ExpandableListView of MenuItem objects from the database.