Completely implemented the ManageMenu module.
Added Java classes include:

AddTablesActivity.java: Provides the user with a form to add a new Table to the database.
ModifyTableActivity.java: Provides the user with a form to modify or delete an existing Table from the database.
TablesBaseAdapter.java: Defines a BaseAdapter for the ListView in the ManageTablesFragment.
TablesValueEventListener: Defines a ValueEventListener used in the ManageTablesFragment.
ManageTablesFragment: Provides the user with a ListView of Table objects from the database.