Partially implemented TableList module.
Added/modified Java classes include:

TableOrderActivity.java: Largely unimplemented. Will eventually have an ExpandableListView of MenuItemsWithQuantity that the user may manipulate as one of the attributes of an Order object they will submit to the database.
TableListBaseAdapter.java: Defines a base adapter used for the ListView in the TableListFragment.
DialogChildEventListener.java: Defines a ChildEventListener that closes dialog boxes when it's code is executed.
TableListFragment.java: A Fragment with a ListView of Table objects. When an object is clicked, a dialog of options starts. Clicking "Update Status" starts another dialog of status options the user may apply to the selected object. Clicking "Take Order" starts the TableOrderActivity, passing the attributes of the clicked object to the activity.

Added/modified layouts include:

activity_table_order.xml: An unfinished layout for the TableOrderActivity.
item_table_employee_view.xml: The view of a single item of the ListView in the TableListFragment.
item_table_manager_view: Was simply renamed as so to distinguish it from the item_table_employee_view.xml layout file. Is the view of a single item of the ListView in the ManageTablesFragment.