Improvements to performance. Added sorting to existing ListViews.
Removed on click listeners from all base adapter classes, as I've learned that this is inefficient. ListView clicks are now handled in the activity class they are used in. So, instead of setting a click listener on each view of the ListView, we are setting a click listener on the entire ListView.

Added sorting to all existing ListViews across the application. Employee, InventoryItem, MenuItem, and Table objects are all sorted by their name attributes.